Move EmbeddingsModel types to types folder
