+++ js/packages/teams-ai/src/embeddings/index.ts
@@ -5,6 +5,4 @@
  * Copyright (c) Microsoft Corporation. All rights reserved.
  * Licensed under the MIT License.
  */
-
-export * from './EmbeddingsModel';
 export * from './OpenAIEmbeddings';
