+++ js/packages/teams-ai/src/types/EmbeddingsModel.ts
